JDBC implementations were moved to separate package
